requirements.txt
openai>=1.12.0
python-dotenv>=1.0.0
requests>=2.31.0

# 向量数据库与嵌入
chromadb>=0.5.3
tiktoken>=0.7.0

# 语言模型与工具链
langchain>=0.2.12
langchain-community>=0.2.10
langchain-core>=0.2.24

# 数据与处理支持
numpy>=1.24.0
pandas>=2.2.0

# 命令行/交互
rich>=13.7.0
colorama>=0.4.6
tqdm>=4.66.0